% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysreqs.R
\name{ci_new_pkgs_sysreqs}
\alias{ci_new_pkgs_sysreqs}
\title{convert a list of packages into a description file and detect system requirements}
\usage{
ci_new_pkgs_sysreqs(pkgs, ...)
}
\arguments{
\item{pkgs}{the list of packages, each an element with a \code{package} field}

\item{...}{extra options to be passed to ci_sysreqs()`}
}
\description{
This function converts a set of R package names to a description file and then
passes this to \code{\link[=ci_sysreqs]{ci_sysreqs()}} for system requirements detection.
}
\examples{
Sys.setenv(R_USER_CACHE_DIR = tempfile())

pkgs <- list(list(package = "knitr", package = "rmarkdown"))

vise::ci_new_pkgs_sysreqs(pkgs, execute = FALSE)
}
