% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysreqs.R
\name{ci_sysreqs}
\alias{ci_sysreqs}
\title{detect and execute system requirements from a lockfile}
\usage{
ci_sysreqs(
  lockfile,
  execute = TRUE,
  sudo = TRUE,
  exclude = c("git", "make", "pandoc")
)
}
\arguments{
\item{lockfile}{the path to a {renv} lockfile}

\item{execute}{if \code{TRUE} (default), the commands from \code{\link[remotes:system_requirements]{remotes::system_requirements()}}
will be executed via \code{\link[base:system]{base::system()}}.}

\item{sudo}{if \code{TRUE} (default), the command runs as root}

\item{exclude}{packages to exclude from installation because they already
exist on the system.}
}
\value{
a vector of exectutible system calls to install the dependencies
}
\description{
This function converts a renv lockfile to a description file and then
determines the system requirements via \code{\link[remotes:system_requirements]{remotes::system_requirements()}} using
the RStudio online resource. This is intended to be run on continuous
integration, so will probably error on macos and windows
}
\examples{
lock <- system.file("renv.lock", package = "vise")

# The system requirements for a typical {knitr} installation
ci_sysreqs(lock, execute = FALSE)
}
