% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_to_f2.R
\name{d_to_f2}
\alias{d_to_f2}
\title{Convert d to Cohen's f}
\usage{
d_to_f2(d)
}
\arguments{
\item{d}{the effect size to convert}
}
\value{
Both Cohen's f and f^2 statistics

\item{f}{d values translated into f}
\item{f2}{d values translated into f^2}
}
\description{
This function allows you to convert d to
Cohen's f and f^2 statistics.
}
\keyword{cohen's}
\keyword{convert}
\keyword{d,}
\keyword{effect}
\keyword{f}
\keyword{f,}
\keyword{size,}
\keyword{squared,}
