% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_to_nnt.R
\name{d_to_nnt}
\alias{d_to_nnt}
\title{Convert d to Number Needed to Treat}
\usage{
d_to_nnt(d = NULL)
}
\arguments{
\item{d}{the effect size}
}
\description{
This function calculates the number needed to treat
from continuous measures (Cohen's d) using Kraemer and Kupfer
(2006) formula.
}
\references{
Kraemer H.C., Kupfer D.J. (2006) Size of treatment effects and their
importance to clinical research and practice.
\emph{Biolological Psychiatry, 59}, 990–996.
https://doi.org/10.1016/j.biopsych.2005.09.014
}
