% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_to_r.R
\name{d_to_r}
\alias{d_to_r}
\title{Convert d to correlation coefficient}
\usage{
d_to_r(d)
}
\arguments{
\item{d}{the effect size to convert}
}
\value{
correlation coefficient
}
\description{
This function allows you to convert d to Pearson's
correlation coefficient.
}
\examples{
d_to_r(.25)

}
\keyword{cohen's}
\keyword{convert,}
\keyword{correlation}
\keyword{d,}
\keyword{effect}
\keyword{size,}
