% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_r.R
\name{estimate_r}
\alias{estimate_r}
\title{Visualization for Estimating \eqn{r}}
\usage{
estimate_r(r = NULL)
}
\arguments{
\item{r}{a correlation to visualize}
}
\value{
Returns a pretty graph

\item{graph}{A graph of the effect size}
}
\description{
This function displays a visualization of effect sizes.
}
\examples{
estimate_r(r = .4)

}
\keyword{effect}
\keyword{estimation}
\keyword{ggplot}
\keyword{size}
\keyword{visualization}
