% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lock2desc.R
\name{lock2desc}
\alias{lock2desc}
\title{convert a lockfile to a description file}
\usage{
lock2desc(lockfile, desc = tempfile())
}
\arguments{
\item{lockfile}{the path to the renv lockfile OR a \code{\link[desc:description]{desc::description()}} object.}

\item{desc}{the path to the new description file}
}
\value{
the path to the new description file
}
\description{
By default, this will take in a lockfile or a desc object and convert it to
an equivalent DESCRIPTION file for use with packages that check for system
dependencies.
}
\examples{
lock <- system.file("renv.lock", package = "vise")
lock2desc(lock)
}
