% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meow.R
\name{meow}
\alias{meow}
\title{\code{cat()} for GitHub output variables}
\usage{
meow(name, thing)
}
\arguments{
\item{thing}{a vector that can be passed to \code{cat()}}

\item{the}{name of the output variable to write to}
}
\value{
nothing. This is used to write output for GitHub
}
\description{
\code{cat()} for GitHub output variables
}
\details{
This is a way to print output for GitHub that requires you to pipe output into
the \code{GITHUB_OUTPUT} environment variable. This will write single-line
variables as \code{var=value} and multiline variables using a random key
according to this guidance from GitHub to avoid injection:
\url{https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#multiline-strings}
}
\examples{
# NOTE: this function is not exported
tmp <- tempfile()
env <- Sys.getenv("GITHUB_OUTPUT", unset = "")
on.exit(Sys.setenv("GITHUB_OUTPUT" = env))
Sys.setenv("GITHUB_OUTPUT" = tmp)
vise <- asNamespace("vise")
vise$meow("single", "hello!")
vise$meow("multi", c("This information", "exists on", "multiple lines"))
writeLines(readLines(tmp))
}
\keyword{internal}
