% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_to_d.R
\name{other_to_d}
\alias{other_to_d}
\title{Convert other statistics to d}
\usage{
other_to_d(
  f = NULL,
  f2 = NULL,
  nnt = NULL,
  r = NULL,
  prob = NULL,
  prop_u1 = NULL,
  prop_u2 = NULL,
  prop_u3 = NULL,
  prop_overlap = NULL
)
}
\arguments{
\item{f}{Cohen's f}

\item{f2}{Cohen's f squared}

\item{nnt}{Number needed to treat}

\item{r}{Correlation coefficient}

\item{prob}{Probability superiority}

\item{prop_u1}{Proportion Overlap U1}

\item{prop_u2}{Proportion Overlap U2}

\item{prop_u3}{Proportion Overlap U3}

\item{prop_overlap}{Proportion Overlap of Distributions}
}
\value{
d
}
\description{
This function allows you to convert other effect sizes
to d including f, f squared, number needed to treat,
correlation coefficient, probability of superiority,
proportion overlap (u1, u2, u3, and proportion distribution
overlap). Please note these are approximations.
}
\examples{
other_to_d(f = .1)

}
\keyword{cohen's}
\keyword{convert,}
\keyword{d,}
\keyword{effect}
\keyword{size,}
