% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_simple_vector.R
\name{verify_simple_vector}
\alias{verify_simple_vector}
\title{Verify that an input is a simple vector}
\usage{
verify_simple_vector(inputs)
}
\arguments{
\item{inputs}{an R expression}
}
\value{
nothing. this is called for its side-effect. if \code{inputs} is not a
simple vector, an error will be thrown.
}
\description{
In some GitHub actions, you can pass an arbitrary R vector in and we will
parse it. Unfortunately, this can lead to the insertion of malicious code,
so this function will thrown an error if the submitted R expression is
anything but a vector created with \code{c()},
}
\examples{
# simple vectors work.
try(vise::verify_simple_vector(c(1, "a", c(TRUE, FALSE))))

# vectors using any other function fail, even if they are harmless
try(vise::verify_simple_vector(c(1, print("a"), c(TRUE, FALSE))))
}
