% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vise-package.R
\docType{package}
\name{vise-package}
\alias{vise}
\alias{vise-package}
\title{vise: Package Discovery and Management Based on 'renv'}
\description{
The 'renv' package is a wonderful tool for managing projects that is extensible and reliable. Vise serves as an opinionated implementation of 'renv' that provides an one-function interface to the hydrate, restore, snapshot workflow.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zkamvar/vise}
  \item Report bugs at \url{https://github.com/zkamvar/vise/issues}
}

}
\author{
\strong{Maintainer}: Zhian N. Kamvar \email{zkamvar@gmail.com} (\href{https://orcid.org/0000-0003-1458-7108}{ORCID})

}
\keyword{internal}
