% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_c.R
\name{visualize_c}
\alias{visualize_c}
\title{Visualization for Estimating c Bias}
\usage{
visualize_c(dlow)
}
\arguments{
\item{dlow}{The lower limit of the possible effect size}
}
\value{
Returns a pretty graph

\item{graph}{The graph of possible values for c}
}
\description{
This function displays a visualization of the possible
bias c that allows for a non-zero effect in sensitivity.
}
\examples{

visualize_c(dlow = .25)

}
\keyword{effect}
\keyword{estimation,}
\keyword{ggplot}
\keyword{size,}
\keyword{visualization,}
