% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_c_map.R
\name{visualize_c_map}
\alias{visualize_c_map}
\title{Visualization for Estimating c Bias + Estimates}
\usage{
visualize_c_map(dlow, dvalues, rvalues)
}
\arguments{
\item{dlow}{The lower limit of the possible effect size}

\item{dvalues}{A vector of effect size values that are
possible.}

\item{rvalues}{A vector of correlation values that are
possible.}
}
\value{
Returns a pretty graph

\item{graph}{The graph of possible values for c}
}
\description{
This function displays a visualization of the possible
bias c that allows for a non-zero effect in sensitivity.
This function includes the ability to add values of
effect size and correlation to see how they map onto
the proposed c value.
}
\examples{

visualize_c_map(dlow = .25,
  dvalues = c(.2, .3, .8),
  rvalues = c(.1, .4, .3))

}
\keyword{effect}
\keyword{estimation,}
\keyword{ggplot}
\keyword{size,}
\keyword{visualization,}
