% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lock2desc.R
\name{lock2desc}
\alias{lock2desc}
\title{convert a lockfile to a description file}
\usage{
lock2desc(lockfile, desc = tempfile())
}
\arguments{
\item{lockfile}{the path to the renv lockfile}

\item{desc}{the path to the new description file}
}
\value{
desc
}
\description{
convert a lockfile to a description file
}
\examples{
lock <- system.file("renv.lock", package = "vise")
lock2desc(lock)
}
