% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_effects.R
\name{visualize_effects}
\alias{visualize_effects}
\title{Visualization for Conversions of Effect Sizes}
\usage{
visualize_effects(d)
}
\arguments{
\item{d}{d effect size to convert to other numbers}
}
\value{
Returns a pretty graph of all the effects

\item{graph}{ggplot object of converted effect sizes}
}
\description{
This function displays a visualization the same effect in
various effect sizes including d, f, \eqn{f^2}, proportion
overlap, correlation, number needed to treat, and more.
}
\examples{

visualize_effects(d = .25)

}
\keyword{effect}
\keyword{estimation,}
\keyword{ggplot}
\keyword{size,}
\keyword{visualization,}
